Fix building wheels for Windows (#81)

+++ sherpa-onnx/python/csrc/features.cc
@@ -11,7 +11,7 @@ namespace sherpa_onnx {
 static void PybindFeatureExtractorConfig(py::module *m) {
   using PyClass = FeatureExtractorConfig;
   py::class_<PyClass>(*m, "FeatureExtractorConfig")
-      .def(py::init<float, int32_t, int32_t>(),
+      .def(py::init<int32_t, int32_t, int32_t>(),
            py::arg("sampling_rate") = 16000, py::arg("feature_dim") = 80,
            py::arg("max_feature_vectors") = -1)
       .def_readwrite("sampling_rate", &PyClass::sampling_rate)
